+++ akm_src/slpi_proc/ssc/sensors/ak0991x/src/sns_ak0991x_sensor_instance.h
@@ -152,7 +152,7 @@ typedef struct ak0991x_self_test_info
 typedef struct ak0991x_config_event_info
 {
   ak0991x_mag_odr  odr;
-  uint32_t         fifo_wmk;
+  uint16_t         fifo_wmk;
   uint32_t         dae_wmk;
 } ak0991x_config_event_info;
